1. added permission for internet 2. added test api call

+++ app/src/main/java/com/example/pranit/mvpshop/data/models/Tax.kt
@@ -10,6 +10,6 @@ class Tax {
     var name: String? = null
     @SerializedName("value")
     @Expose
-    var value: Int? = null
+    var value: Float? = null
 
 }
